+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRZW3QH57E440NYH5JVX684V"
+        "01KS04FBBCWERWKYFC1ZD1XTWC"
     )
     r"""The unique identifier of the data source"""