+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRZW3QH8H8QER8FH46RCZCGA"
+        "01KS04FBBFW0T156QF1G4EG60Z"
     )
     r"""The unique identifier of the data source"""