+++ packages/orq-rc/RELEASES.md
@@ -8019,3 +8019,13 @@ Based on:
 - [python v4.10.0-rc.16] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.16 - packages/orq-rc
+
+## 2026-05-19 12:51:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.1 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.17] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.17 - packages/orq-rc